FIX: Iteration and mutation of primaries_down in separate threads
Ruby hashes can't be modified whilst they are being iterated over.

Here, the primaries_down hash is iterated over to check each previously
unavailable primary to see if it is now contactable. However, since this
hash can be updated in other threads, this iteration isn't safe.

To prevent this, a copy of the hash is iterated over instead.

The GIL should not be released during a hash dup [1], but let's not tie
ourselves unnecessarily to current MRI behaviour.

[1]: https://github.com/ruby-concurrency/concurrent-ruby/issues/528