FIX: Force redis disconnect if lock acquisition takes too long
Previously we were waiting to acquire the mutex for each connection before disconnecting it. This tries to ensure a clean disconnection, but has two issues:

- It will block forever on long-running redis commands (e.g. Pub/Sub subscriptions)
- When two clients trigger failover simultaneously, it can cause a deadlock where they each wait for the other's mutex

This commit adds a 50ms timeout, after which the client is terminated forcefully. This may cause some form of Redis::BaseConnectionError, but we expect to see some errors like this during failover anyway, so consumers should be tolerant to it.

This commit also fixes the disconnection for redis SubscribedClient instances, which were previously being skipped.