FIX: Various potential concurrency issues (#15)
This uses concurrent-ruby maps to simplify concurrency logic. Resolves a number of possible concurrency issues

- Removes ancestor_pid instance variable, and now supports multiple forks
- Avoids calling the failover/fallback callbacks from within a mutex
- Avoids iterating over native ruby hashes
- Removes potential race conditions like `@primaries_down[process_pid] if @primaries_down[process_pid]`, where the value could change between each read

The Concurrent::Map api provides a number of helpers which help to reduce the amount of code requires for thread safety.

concurrent-ruby was already a transitive dependency, so this does not materially change the gem's dependencies.