FIX: Handle the case when the replica is set equal to the primary (#20)
ed8634cc introduced a check which closed connections which do not match the expected role. In the case where the replica and primary were the same, it identified every connection as the replica, and closed it immediately. This commit flips the logic so it will identify the connection as the primary, and allow it to open successfully.