DEV: Make ActiveRecord tests more stable
This commit updates the ActiveRecord tests to be more reliable by
reducing the Unicorn worker processes to 1 so that we don't have to rely
on flooding the app with requests to get all the Unicorn processes to
failover/fallback.

The commit also updates the way we start/stop the Rails and Postgres
servers to ensure that every RSpec tests starts from the same consistent
state. This was not the case before this commit.