FIX: Falling back to primary PG server not reliable on Rails 7.1
This commit drops the reliance on
`ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?` in favor
or using `ActiveRecord::ConnectionAdatpers::PostgreSQLAdatper#execute`
to check if the app can connect to the primary PG server.

This is because `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?`
was changed in Rails 7.1 to return `false` if the connection has not be used to do
something meaningful. Ref:
https://github.com/rails/rails/commit/8551e64e2411811f26d210601abdba6e13d8798c.
Due to this change, our fallback to primary checker will keep thinking
that the primary PG server is down. Note that before Rails 7.1, `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?`
will always execute a fake query to check if the query can be executed.

Instead of relying on ActiveRecord's internal API, we will instead rely
on `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#execute` to
execute a blank query as our means of verifying if the PG server is up
and ready to receive connections.

This commit also updates the ActiveRecord tests to be more reliable by
reducing the Unicorn worker processes to 1 so that we don't have to rely
on flooding the app with requests to get all the Unicorn processes to
failover/fallback.

Co-authored-by: Loïc Guitaut <loic@discourse.org>