FIX: Falling back to primary PG server not reliable on Rails 7.1 (#71)
This commit adds the use of [ActiveRecord::ConnectionAdatpers::PostgreSQLAdatper#verify!](https://api.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/AbstractAdapter.html#method-i-verify-21)
when a new connection is retrieved from the connection pool. This is because `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?`
was changed in Rails 7.1 to return `false` if the connection has not be used to do
something meaningful. Ref:
https://github.com/rails/rails/commit/8551e64e2411811f26d210601abdba6e13d8798c.
Due to this change, our fallback to primary checker will keep thinking
that the primary PG server is down. Note that before Rails 7.1, `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?`
will always execute a fake query to check if the query can be executed. By calling `ActiveRecord::ConnectionAdatpers::PostgreSQLAdatper#verify!`, before `ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#active?`, we ensure that the connection to the Postgres 
server has been made first.

This commit also updates the ActiveRecord tests to be more reliable by
reducing the Unicorn worker processes to 1 so that we don't have to rely
on flooding the app with requests to get all the Unicorn processes to
failover/fallback.

Co-authored-by: Loïc Guitaut <loic@discourse.org>