DEV: Fix flaky tests on CI (#74)
`GenericHelper#wait_for` implementation was incorrect as the timeout
parameter is in seconds but the code processes it in millieseconds.

Also make creating and dropping the test database more reliable by managing it in PG directly. 
Also improve the various start/stop PG primary and replica servers command to check for PG's status.